+++ packages/media-fields/src/author/index.tsx
@@ -16,7 +16,7 @@ const authorField: Partial< Field< MediaItem > > = {
 	type: 'integer',
 	getElements: async () => {
 		const authors: Author[] =
-			( await resolveSelect( coreDataStore ).getEntityRecords(
+			( await resolveSelect( coreDataStore ).getEntityRecords< Author >(
 				'root',
 				'user',
 				{